pyproject.toml
[tool.poetry]
name = "fastpass"
version = "0.1.0"
description = ""
authors = ["Link13-tech <151656048+Link13-tech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
